remove some 3D elem from home page
