+++ src/assets/scss/variables.scss
@@ -119,7 +119,7 @@
 
     --fs-heading: clamp(45px, 5vw, 90px);
 
-    --fs-text-banner: clamp(25px, 2.3vw, 33px);
+    --fs-text-banner: clamp(25px, 2.2vw, 36px);
     --fs-text-banner-sm: clamp(22px, 1.8vw, var(--fs-lg));
 
     @include mixins.mobile-tablet-vertical {
