+++ src/components/Header/Header.module.scss
@@ -17,7 +17,7 @@
 
     position: fixed;
     width: 100vw;
-    top: 20px;
+    top: 5px;
     z-index: 1000;
 
     &[data-theme="dark"] .list:first-child a {
