+++ src/components/TextBanner/TextBanner.module.scss
@@ -72,7 +72,8 @@
 .text {
     position: relative;
     margin: 0;
-    color: var(--c-font-global);
+    color: var(--c-font-global-light);
+    font-family: "inter-medium";
     font-size: var(--fs-text-banner);
     font-weight: var(--fw-regular);
     z-index: 10;
